sound/codecs: Fix for merlin
There are two cs35l41 drivers in the code, the first driver is from ALPS and the second driver is from merlin-r-oss. The cs35l41 driver is used only by merlin, so it is required to leave only the driver from merlin-r-oss. I also replaced the SND_SOC_WM_ADSP driver from ALPS with the variant from merlin-r-oss.